Creacion de la clase tag y conexion de la interfaz oferta con el backend para poder crear ofertas desde el

+++ SpringBoot/src/main/java/com/emprendev/services/OfferServices.java
@@ -1,10 +1,12 @@
 package com.emprendev.services;
+
 import com.emprendev.exceptions.ResourceNotFoundException;
 import com.emprendev.entity.Offer;
 import com.emprendev.repository.OfferRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,17 +25,33 @@ public class OfferServices {
     }
 
     public Offer saveOffer(Offer offer) {
+        // Establecer valores predeterminados
+        if (offer.getCreationDate() == null) {
+            offer.setCreationDate(String.valueOf(LocalDate.now())); // Fecha actual
+        }
+        if (offer.getFinalizationDate() == null) {
+            offer.setFinalizationDate(String.valueOf(LocalDate.now().plusMonths(1))); // Fecha dentro de un mes
+        }
+        if (offer.getOfferState() == 0) {
+            offer.setOfferState(1); // Estado inicial
+        }
         return offerRepository.save(offer);
     }
 
     public void deleteOffer(Long id) {
-        offerRepository.deleteById(id);
+        if (offerRepository.existsById(id)) {
+            offerRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Offer not found with id " + id);
+        }
     }
 
     public Offer updateOffer(Long id, Offer offerDetails) {
         Optional<Offer> optionalOffer = offerRepository.findById(id);
         if (optionalOffer.isPresent()) {
             Offer offer = optionalOffer.get();
+
+            // Actualizar los detalles de la oferta
             offer.setTitle(offerDetails.getTitle());
             offer.setDescription(offerDetails.getDescription());
             offer.setCreationDate(offerDetails.getCreationDate());
@@ -45,10 +63,13 @@ public class OfferServices {
             offer.setImageUrl3(offerDetails.getImageUrl3());
             offer.setImageUrl4(offerDetails.getImageUrl4());
             offer.setOfferState(offerDetails.getOfferState());
+
+            // Actualizar los tags
+            offer.setTags(offerDetails.getTags());
+
             return offerRepository.save(offer);
         } else {
-            throw new ResourceNotFoundException("Offer not found with id juajuas" + id);
+            throw new ResourceNotFoundException("Offer not found with id " + id);
         }
     }
 }
-
